+++ misc/dbt-materialize/dbt/adapters/materialize/connections.py
@@ -104,28 +104,25 @@ class MaterializeConnectionManager(PostgresConnectionManager):
         connection.handle.autocommit = True
         return connection
 
+    # Disable transactions. Materialize transactions do not support arbitrary
+    # queries in transactions and therefore many of dbt's internal macros
+    # produce invalid transactions.
+    #
+    # Disabling transactions has precedent in dbt-snowflake and dbt-biquery.
+    # See, for example:
+    # https://github.com/dbt-labs/dbt-snowflake/blob/ffbb05391/dbt/adapters/snowflake/connections.py#L359-L374
+
+    def add_begin_query(self, *args, **kwargs):
+        pass
+
+    def add_commit_query(self, *args, **kwargs):
+        pass
+
+    def begin(self):
+        pass
+
     def commit(self):
-        connection = self.get_thread_connection()
-        if flags.STRICT_MODE:
-            if not isinstance(connection, Connection):
-                raise dbt.exceptions.CompilerException(
-                    f"In commit, got {connection} - not a Connection!"
-                )
+        pass
 
-        # Instead of throwing an error, quietly log if something tries to commit
-        # without an open transaction.
-        # This is needed because the dbt-adapter-tests commit after executing SQL,
-        # but Materialize can't handle all of the required transactions.
-        # https://github.com/fishtown-analytics/dbt/blob/42a85ac39f34b058678fd0c03ff8e8d2835d2808/test/integration/base.py#L681
-        if not connection.transaction_open:
-            logger.debug(
-                'Tried to commit without a transaction on connection "%s"',
-                connection.name,
-            )
-
-        logger.debug("On %s: COMMIT", connection.name)
-        self.add_commit_query()
-
-        connection.transaction_open = False
-
-        return connection
+    def clear_transaction(self):
+        pass